// src/api/@types/general.types.ts
export enum DepartamentRolesTypes {
	SAC = 'SAC',
	EVENTS = 'Eventos',
	PRODUCTIVE = 'Produtivo',
	QUALITY_ASSURANCE = 'Garantia de qualidade',
	COMERCIAL = 'Comercial',
	MARKETING = 'Marketing',
	MEDICAL_VISITATION = 'Visitação Médica',
	ADMINISTRATIVE = 'Administrativo',
}

export enum NameRolesTypes {
	SUPERVISOR = 'Supervisor',
	MANAGER = 'Gerente',
	ANALYST = 'Analista',
	DIRECTOR = 'Diretor',
}

export enum AgentStatusTypes {
	ACTIVE = 'active',
	INACTIVE = 'inactive',
}